Move protocol and extension into it's own folder
