+++ Source/ViewModels/Attachment.swift
@@ -14,15 +14,3 @@ public struct Attachment {
     self.source = source
   }
 }
-
-public protocol URLStringConvertible {
-  var url: NSURL { get }
-}
-
-extension String: URLStringConvertible {
-
-  public var url: NSURL {
-    return NSURL(string: self)!
-  }
-
-}
